Fix a warning. Cocoapods updated to 1.0.1

+++ DMRESTRequest/DMRESTRequest/DMJSONCache.h
@@ -27,5 +27,5 @@
 //Completly remove in memory cache
 + (void)emptyInMemoryCache;
 //Size for the cache folder
-+ (int)diskCacheSize;
++ (unsigned long long)diskCacheSize;
 @end
